refactor(threads): move listener to its own action
